[DM-15396] Add setup.py for metaserv.
This does a couple of things.  First, use requirements.txt to include
the requirements for the project, rather than a list in setup.py,
because github will help us by looking for requirements.txt files.

Next, add a setup.py file that handles versioning and generating
the package version string, as well as the normal python stuff.

Create a version.py that exposes the version the same way it used to
be.  By checking in a file here, I believe this will make it back
compatible to the old way of using this package, at least for now.